Allow interaction with oc login
`oc login` may ask about skipping tls verification. Allowing the
interaction require using the current process stdin and stdout so oc
login can write prompts and get the response from the user.

Since oc writes errors to both stdout and stderr, it make sense to let
it use the current process stderr as well, so this fixes also the issue
with hidden stderr.

`oc logout` does not need interaction, but capturing both stdout and
stderr ensures we will not miss error output in case it writes it to
stdout.

Fixes: #4
Fixes: #5